Replace card 5 f(x)=sin(x) with -x^3
Card 5 previously used the function sin(x).
We change card 5 to use the function -x^3/3+0.5.
We also change the corresponding derivative card to f'(x)=-x^2.

Students tend to first try to match the f(x) cards with f'(x)
cards that look similar. The fact that f=sin and f'=cos match
only exacerbates this tendency.

There is already a card f(x)=x^3/3+1.
We add this new card f(x)=-x^3/3+0.5 so that students can make
a connection (informally) along the lines of (-f)' = -f'.
We use -x^3/3+0.5 instead of just -x^3 so that students can
make the connection that vertical translation of f "doesn't matter"
in the sense that vertical translation of f does not affect f'.

As a reminder, we use x^3/3 instead of x^3 to make the graphs
clearer. Note: students only see the graphs of f, f', not the formulas.